+++ src/features/User/index.js
@@ -25,7 +25,6 @@ import PvpGame from './components/PvpGame';
 import PvpLeague from './components/PvpSeason';
 import FavouritePvpClass from './components/FavouritePvpClass';
 
-
 import {
   fetchUser,
   selectUser,
@@ -104,10 +103,7 @@ export default class User extends Component {
     return (
       <Content
         cardExtra={user && user.access && (
-          <div className={styles.access}>
-            <Icon size="mini" name={`${user.access}.png`} />
-            <span>{startCase(user.access)}</span>
-          </div>
+          <Icon size="mini" className={styles.access} name={`${user.access}.png`} />
         )}
         type="users"
         title={alias}